// src/generated/models/Fc_campaignmetricsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Fc_campaignmetricsfc_channel = {
  122370000: 'facebook',
  122370001: 'instagram',
  122370002: 'linkedin',
  122370003: 'tiktok',
  122370004: 'x',
  122370005: 'whatsapp',
  122370006: 'google'
} as const;
export type Fc_campaignmetricsfc_channel = keyof typeof Fc_campaignmetricsfc_channel;
export const Fc_campaignmetricsfc_source = {
  122370000: 'manual',
  122370001: 'flowcom',
  122370002: 'buffer',
  122370003: 'odoo',
  122370004: 'whatsapp',
  122370005: 'meta_ads',
  122370006: 'google_ads',
  122370007: 'linkedin_ads'
} as const;
export type Fc_campaignmetricsfc_source = keyof typeof Fc_campaignmetricsfc_source;
export const Fc_campaignmetricsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Fc_campaignmetricsstatecode = keyof typeof Fc_campaignmetricsstatecode;
export const Fc_campaignmetricsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Fc_campaignmetricsstatuscode = keyof typeof Fc_campaignmetricsstatuscode;

export interface Fc_campaignmetricsBase {
  "fc_Campaign@odata.bind": string;
  fc_campaignmetricid: string;
  fc_channel: Fc_campaignmetricsfc_channel;
  fc_clicks?: number;
  "fc_Company@odata.bind": string;
  fc_conversions?: number;
  fc_date: string;
  fc_engagements?: number;
  fc_impressions?: number;
  fc_leads?: number;
  fc_name?: string;
  fc_reach?: number;
  fc_revenue?: number;
  fc_source: Fc_campaignmetricsfc_source;
  fc_spend?: number;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  statecode: Fc_campaignmetricsstatecode;
  statuscode?: Fc_campaignmetricsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Fc_campaignmetrics extends Fc_campaignmetricsBase {
  createdon?: string;
  fc_channelname?: string;
  fc_sourcename?: string;
  modifiedon?: string;
  ownerid: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  fc_campaign?: object;
  _fc_campaign_value?: string;
  fc_company?: object;
  _fc_company_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}